tech documentation submitted
in a practical sense, going through the technical details and comparing Rust to C code isn't really necessary.
I would still like to learn that myself, but for now, having documentation going through the program step by step is sufficient (in my opinion).
The CSS for the buttons needs fixing, and the positioning for containers needs to be changed from pixels to percentages to account for different resolutions.